fix: 하이고 문제 번호 착각
